[refactor/#63] reservation-history 모달 컴포넌트 구조 이동
